+++ README.md
@@ -82,6 +82,6 @@
 * MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현한다.
 
 ### 체크 리스트
-* [ ] Domain 패키지로 이동
+* [x] Domain 패키지로 이동
 * [ ] View 패키지로 이동
 * [ ] View -> Domain 으로만 의존하도록 변경
