+++ src/main/kotlin/car/domain/CarRacing.kt
@@ -1,4 +1,4 @@
-package car.racing
+package car.domain
 
 object CarRacing {
