+++ src/main/kotlin/car/domain/Record.kt
@@ -1,3 +1,3 @@
-package car.racing
+package car.domain
 
 data class Record(val driverName: String, val moves: Int)
