+++ src/test/kotlin/car/domain/CarConditionTest.kt
@@ -1,5 +1,6 @@
-package car.racing
+package car.domain
 
+import car.domain.CarCondition
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
