+++ src/test/kotlin/car/domain/CarRacingTest.kt
@@ -1,5 +1,6 @@
-package car.racing
+package car.domain
 
+import car.domain.CarRacing
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldNotBe
