+++ src/test/kotlin/car/domain/CarsTest.kt
@@ -1,4 +1,4 @@
-package car.racing
+package car.domain
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
