+++ src/test/kotlin/car/domain/GameTest.kt
@@ -1,4 +1,4 @@
-package car.racing
+package car.domain
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FreeSpec
