+++ src/test/kotlin/car/racing/ResultViewTest.kt
@@ -1,5 +1,7 @@
 package car.racing
 
+import car.domain.Cars
+import car.domain.GameRecords
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.ints.shouldBeGreaterThanOrEqual
 import io.kotest.matchers.shouldBe
